Reorganize code and decorate span with event data
