+++ data-plane/receiver/src/main/java/dev/knative/eventing/kafka/broker/receiver/CloudEventRequestToRecordMapper.java
@@ -19,10 +19,13 @@ package dev.knative.eventing.kafka.broker.receiver;
 import static net.logstash.logback.argument.StructuredArguments.keyValue;
 
 import dev.knative.eventing.kafka.broker.core.cloudevents.PartitionKey;
+import dev.knative.eventing.kafka.broker.core.tracing.Tracing;
+import dev.knative.eventing.kafka.broker.core.tracing.TracingSpan;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.message.MessageReader;
 import io.cloudevents.http.vertx.VertxMessageFactory;
 import io.vertx.core.Future;
+import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.kafka.client.producer.KafkaProducerRecord;
 import org.slf4j.Logger;
@@ -32,20 +35,37 @@ public class CloudEventRequestToRecordMapper implements RequestToRecordMapper<St
 
   private static final Logger logger = LoggerFactory.getLogger(CloudEventRequestToRecordMapper.class);
 
+  private final Vertx vertx;
+
+  public CloudEventRequestToRecordMapper(final Vertx vertx) {
+    this.vertx = vertx;
+  }
+
   @Override
   public Future<KafkaProducerRecord<String, CloudEvent>> recordFromRequest(
     final HttpServerRequest request,
     final String topic) {
 
     return VertxMessageFactory.createReader(request)
-      // TODO is this conversion really necessary?
-      //      can be used Message?
       .map(MessageReader::toEvent)
       .map(event -> {
         if (event == null) {
           throw new IllegalArgumentException("event cannot be null");
         }
-        logger.debug("received event {}", keyValue("event", event));
+
+        if (logger.isDebugEnabled()) {
+          final var span = TracingSpan.getCurrent(vertx);
+          if (span != null) {
+            logger.debug("received event {} {}",
+              keyValue("event", event),
+              keyValue(Tracing.TRACE_ID_KEY, span.getSpanContext().getTraceIdAsHexString())
+            );
+          } else {
+            logger.debug("received event {}", keyValue("event", event));
+          }
+        }
+
+        TracingSpan.decorateCurrent(vertx, event);
 
         return KafkaProducerRecord.create(topic, PartitionKey.extract(event), event);
       });
